Release 13/11 (#73)
* [CBTC-61] Remove .gitkeep

* [CBTC-61] Update schema & init data sql & entity

* [CBTC-61] Implement update seat & validation

* [CBTC-61] Revert init data sql & update docker-compose

* [CBTC-61] Remove .gitkeep

* [CBTC-61] Update schema & init data sql & entity

* [CBTC-61] Implement update seat & validation

* [CBTC-61] Revert init data sql & update docker-compose

* [CBTC-62] Update SQL for seat with price & seat_type_id

* [CBTC-62] Update SQL for screening & screening-seat

* [CBTC-62] Update seed data for screening & screening-seat

* implement upload image using R2 Object of cloudflare

* [CBTC-62] Update entity & schema

* [CBTC-62] Implement bad request handle exception

* [CBTC-62] Implement Booking Process

* [CBTC-62] Update SQL for seat with price & seat_type_id

* [CBTC-62] Update SQL for screening & screening-seat

* [CBTC-62] Update seed data for screening & screening-seat

* [CBTC-62] Update entity & schema

* [CBTC-62] Implement bad request handle exception

* [CBTC-62] Implement Booking Process

* [CBTC-62] Update docker compose for initial data

* [CBTC-62] Implement Cancel Booking Process & Fix Entity

* adjusted endpoint upload image

* adjusted as request add api for get minimal element

* [CBTC-99] Implement Seat Layout Service for Hall

* enhance create and update Screening

* [CBTC-101] Fix using class instead of primitive type in PageableDTO

* [CBTC-101] Implement Get Minimal List for Cinema & District, able to filter by name

* enhance api Screening and add field for data

* [CBTC-101] Update Branch Decimal Column Entity & Schema sql

* [CBTC-101] Update Branch Decimal Column Entity & Schema sql

* [CBTC-101] Update schema type & entity branch

* [CBTC-101] Update schema type & entity branch - Part 2

* refactor requests file

* Change API user update by pathVariable

* enhance get uses filter

* enhance get branches filter

* enhance get movies filter

* [CBTC-103] Update ID for Hall Criteria

* [CBTC-103] Update Branch for get all Hall

* enhance get screening filter

* adjust movie filter

* [CBTC-103] Update ID for Hall Criteria

* [CBTC-103] Update Branch for get all Hall

* [CBTC-103] Enhance Get Detail Hall with Seat

* implement Jwt security

* setup authenticate service

* [CBTC-103] Enhance Select Query for Join Entity

* [CBTC-103] Enhance Base Query for various method using join relation

* [CBTC-103] Update fetch type lazy for many to one column

* [CBTC-103] Update env properties & application.yml

* [CBTC-103] Move minimal criteria to new package

* [CBTC-103] Add jooq & logging

* setup config authenticate and authorization

* setup security config for user controller

* adjust permission  user controller

* update user table with encode password (12345678)

* [CBTC-103] Update pagination packae

* [CBTC-103] Update jooq plugin

* [CBTC-103] Enhance District Minimal using Jooq

* implement Jwt security

* setup authenticate service

* setup config authenticate and authorization

* setup security config for user controller

* adjust permission  user controller

* update user table with encode password (12345678)

* [CBTC-103] Add jooq & logging

* [CBTC-103] Remove pageable from district

* [CBTC-103] Update configuration maven

* [CBTC-103] Update directory path

* [CBTC-103] Enhance District Minimal API with search name

* [CBTC-103] Enhance Minimal Cinema API using JooQ

* [CBTC-103] Enhance Minimal Cinema Branch using JooQ & Update package Minimal

* [CBTC-103] Enhance criteria minimal field name -> search

* [CBTC-103] Rename movie_name -> name entity & schema

* [CBTC-103] Enhance API minimal movie

* [CBTC-103] Enhance API Screening Minimal

* [CBTC-103] Update path for maven configuration jooq

* [CBTC-103] Implement Get Minimal Provinces API

* [CBTC-103] Change field movieName -> name

* [CBTC-103] Change field movieName -> name (part 2)

* [CBTC-99] Implement Seat Layout Service for Hall

* [CBTC-99] Implement Seat Layout Service for Hall

* [CBTC-103] Implement API Get Minimal Halls

* [CBTC-103] Update Merge condition districts & province in Districts Minimal

* [CBTC-103] Add time property to get minimal movies

* [CBTC-99] Implement Seat Layout Service for Hall

* [CBTC-99] Fix: Validation reservation time & Get seat with screeningId

* [CBTC-99] Reformat code

* [CBTC-99] Fix bugs override seat_screening when hallId equal

* [CBTC-99] Remove unnecessary screening response dto

* [CBTC-99] Delete Hall throw error when having screening

* [CBTC-99] Allow changed hall when screening created & create/update seat through update screening API & Add validation when delete hall

* handle jwt and security exception

* [CBTC-99] Enhance Get Seat Layout for Screening

* [CBTC-99] Update readme

* [CBTC-99] Bypass seatType due to mapper

* [CBTC-99] Update application properties write_dates_as_timestamps

* [CBTC-99] Update application properties write_dates_as_timestamps (2)

* Fix format localDateTime response

* Validate for user creation and update request

* Implement validation creation and update request

* [CBTC-99] Screening Status Match in order to update seats & Change status of screening when user booking

* [CBTC-99] Update Minimal Screening Enum & Validation Start Time

* [CBTC-99] Update string to enum in delete screening

* [CBTC-99] Return screening status in screening list

* update validation request

* [CBTC-113] Enhance JWT Authentication & Authorization

* Implement ProfileController

* Invalidate request as required

* [CBTC-90] Implement Get Statistic

* [CBTC-90] Enhance Get Statistic Screening getting hall name, hall branch address

* [CBTC-90] Enhance Get Statistic Screening getting hall name, hall branch address

* [CBTC-90] Update DTO for Statistic

---------

+++ src/main/java/org/cybersoft/bookingticketcinemabe/service/impl/HallServiceImpl.java
@@ -4,14 +4,16 @@ import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.cybersoft.bookingticketcinemabe.dto.PageableDTO;
 import org.cybersoft.bookingticketcinemabe.dto.hall.HallDetailDTO;
+import org.cybersoft.bookingticketcinemabe.dto.hall.HallDetailSeatDTO;
 import org.cybersoft.bookingticketcinemabe.dto.hall.HallDetailSeatLayoutDTO;
 import org.cybersoft.bookingticketcinemabe.entity.BranchEntity;
 import org.cybersoft.bookingticketcinemabe.entity.HallEntity;
 import org.cybersoft.bookingticketcinemabe.entity.HallEntity_;
 import org.cybersoft.bookingticketcinemabe.entity.SeatEntity;
 import org.cybersoft.bookingticketcinemabe.exception.NotFoundException;
+import org.cybersoft.bookingticketcinemabe.exception.runtime.BadRequestException;
 import org.cybersoft.bookingticketcinemabe.mapper.HallMapper;
-import org.cybersoft.bookingticketcinemabe.mapper.PageableMapper;
+import org.cybersoft.bookingticketcinemabe.mapper.pagination.PageableMapper;
 import org.cybersoft.bookingticketcinemabe.payload.request.hall.*;
 import org.cybersoft.bookingticketcinemabe.query.CriteriaApiHelper;
 import org.cybersoft.bookingticketcinemabe.query.dto.Pageable;
@@ -20,9 +22,9 @@ import org.cybersoft.bookingticketcinemabe.repository.BranchRepository;
 import org.cybersoft.bookingticketcinemabe.repository.HallRepository;
 import org.cybersoft.bookingticketcinemabe.repository.SeatRepository;
 import org.cybersoft.bookingticketcinemabe.service.HallService;
+import org.cybersoft.bookingticketcinemabe.service.SeatLayoutService;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -39,6 +41,8 @@ public class HallServiceImpl implements HallService {
 
     private final HallMapper hallMapper;
 
+    private final SeatLayoutService seatLayoutService;
+
     @Override
     public PageableDTO<?> getHalls(HallCriteria hallCriteria) {
         Pageable pageable = Pageable.builder()
@@ -48,6 +52,10 @@ public class HallServiceImpl implements HallService {
                 .build();
 
         SelectQueryImpl<HallEntity> halls = this.criteriaApiHelper.select(HallEntity.class);
+
+        if (hallCriteria.getId() != null) {
+            halls.equal(HallEntity_.id, hallCriteria.getId());
+        }
 
         if (hallCriteria.getName() != null) {
             halls.like(HallEntity_.name, hallCriteria.getName());
@@ -72,10 +80,15 @@ public class HallServiceImpl implements HallService {
     }
 
     @Override
-    public HallDetailDTO getHall(Integer id) {
+    public HallDetailSeatDTO getHall(Integer id) {
         HallEntity hall = this.hallRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Hall " + id + " not found"));
-        return this.hallMapper.toHallDetailDto(hall);
+        HallDetailSeatDTO hallDetailSeatDTO = this.hallMapper.toHallDetailSeatDTO(hall);
+
+        // TODO: Enhance using mapper
+        hallDetailSeatDTO.setSeats(this.getSeatsLayout(id));
+
+        return hallDetailSeatDTO;
     }
 
     @Transactional
@@ -132,6 +145,11 @@ public class HallServiceImpl implements HallService {
         HallEntity hall = this.hallRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Hall " + id + " not found"));
 
+        if (!hall.getScreenings()
+                .isEmpty()) {
+            throw new BadRequestException("Cannot delete hall " + id + " due to screenings");
+        }
+
         List<SeatEntity> seats = this.seatRepository.findAllByHallId(id);
         this.seatRepository.deleteAllInBatch(seats);
 
@@ -140,24 +158,7 @@ public class HallServiceImpl implements HallService {
 
     @Override
     public List<List<HallDetailSeatLayoutDTO>> getSeatsLayout(Integer id) {
-        List<List<HallDetailSeatLayoutDTO>> seatsLayout = new ArrayList<>();
-
-        List<SeatEntity> seats = this.seatRepository.findAllByHallId(id);
-        seats.forEach(seat -> {
-            HallDetailSeatLayoutDTO seatLayout = this.hallMapper.toHallDetailSeatLayoutDto(seat);
-
-            int seatRow = seatLayout.getSeatRow()
-                                  .charAt(0) - 'A';
-            int seatColumn = seatLayout.getSeatColumn() - 1;
-
-            if (seatColumn == 0) {
-                seatsLayout.add(seatRow, new ArrayList<>());
-            }
-            seatsLayout.get(seatRow)
-                    .add(seatColumn, seatLayout);
-        });
-
-        return seatsLayout;
+        return this.seatLayoutService.getSeatLayoutByHallId(id);
     }
 
     private List<SeatEntity> createSeats(HallEntity hall, List<HallCreateSeat> seats) {
